% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucssr.R
\name{ucssr_open_file}
\alias{ucssr_open_file}
\alias{ucssr_browse}
\title{Open slides, or their Rmarkdown files.}
\usage{
ucssr_open_file(vignette = c("index.Rmd", "corpora.Rmd", "subcorpora.Rmd",
  "counting.Rmd", "kwic.Rmd", "cqp.Rmd", "sentiments.Rmd"))

ucssr_browse(vignette = c("index.html", "corpora.html", "subcorpora.html",
  "counting.html", "kwic.html", "cqp.html", "sentiments.html"))
}
\arguments{
\item{vignette}{The vignette to open, either}
}
\description{
Open slides, or their Rmarkdown files.
}
\examples{
\dontrun{
ucssr_open_file("kwic")
}
\dontrun{
ucssr_browse("cqp")
}
}
