% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucssr.R
\name{open_rmarkdown}
\alias{open_rmarkdown}
\alias{slides}
\title{Open slides, or their Rmarkdown files.}
\usage{
open_rmarkdown(topic = c("index.Rmd", "corpora.Rmd", "subcorpora.Rmd",
  "counting.Rmd", "kwic.Rmd", "cqp.Rmd", "sentiments.Rmd"))

slides(topic = c("index.html", "corpora.html", "subcorpora.html",
  "counting.html", "kwic.html", "cqp.html", "sentiments.html"))
}
\arguments{
\item{topic}{A character string giving the name of the vignette to
view. If omitted, all vignettes from all installed packages are listed.}
}
\description{
The package includes a set of slides that introduce the functionality of
the polmineR package, including workflows and recipes to address specific
typical problems.
}
\examples{
\dontrun{
open_rmarkdown("kwic")
}
\dontrun{
slides("cqp")
}
}
